backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
gunicorn
whitenoise
psycopg2-binary
PyMySQL
dj-database-url
